Clean up code and structure, Added addScore function and counter for attemps
