+++ src/main/java/nl/hu/cisq1/lingo/domain/Feedback.java
@@ -2,8 +2,7 @@ package nl.hu.cisq1.lingo.domain;
 
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
-import nl.hu.cisq1.lingo.domain.exception.InvalidFeedbackException;
-
+import nl.hu.cisq1.lingo.domain.exception.FeedbackInvalidException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,7 +14,7 @@ public class Feedback {
 
     public Feedback(String attempt, List<Mark> marks) {
         if (attempt.length() != marks.size()) {
-            throw new InvalidFeedbackException();
+            throw new FeedbackInvalidException();
         }
 
         this.attempt = attempt;
