+++ src/main/java/nl/hu/cisq1/lingo/domain/Round.java
@@ -1,28 +1,30 @@
 package nl.hu.cisq1.lingo.domain;
 
 import lombok.EqualsAndHashCode;
+import lombok.Getter;
 import lombok.ToString;
-import nl.hu.cisq1.lingo.words.domain.Word;
+import nl.hu.cisq1.lingo.domain.exception.RoundAttemptLimitException;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @EqualsAndHashCode
 @ToString
+@Getter
 public class Round {
-    private int number;
-    private Word word;
+    private final int number;
+    private final Word word;
+    private int attempt;
     private List<Feedback> feedbacks;
 
     public Round(int number, Word word, List<Feedback> feedbacks) {
         this.number = number;
         this.word = word;
+        this.attempt = 0;
         this.feedbacks = feedbacks;
     }
 
     public List<Character> startRound() {
-        this.number ++;
-
         List<Character> characters = new ArrayList<>();
 
         for (int i=0; i< this.word.getLength(); i++) {
@@ -31,16 +33,18 @@ public class Round {
             if(i == 0) {
                 characters.add(letter);
             } else {
-                characters.add('_');
+                characters.add('.');
             }
         }
 
         return characters;
     }
 
-
     public Feedback guessWord(String attempt) {
-        this.number ++;
+
+        if(this.attempt >=5) {
+            throw new RoundAttemptLimitException();
+        }
 
         List<Mark> marks = new ArrayList<>();
 
@@ -55,8 +59,8 @@ public class Round {
             }
         }
 
-        Feedback feedback = new Feedback(attempt, marks);
+        this.attempt++;
 
-        return feedback;
+        return new Feedback(attempt, marks);
     }
 }
