+++ src/main/java/nl/hu/cisq1/lingo/domain/Word.java
@@ -1,0 +1,27 @@
+package nl.hu.cisq1.lingo.domain;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+
+@Entity(name = "words")
+public class Word {
+    @Id
+    @Column(name = "word")
+    private String value;
+    private Integer length;
+
+    public Word() {}
+    public Word(String word) {
+        this.value = word;
+        this.length = word.length();
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public Integer getLength() {
+        return length;
+    }
+}
