+++ src/main/java/nl/hu/cisq1/lingo/domain/exception/FeedbackInvalidException.java
@@ -1,7 +1,7 @@
 package nl.hu.cisq1.lingo.domain.exception;
 
-public class InvalidFeedbackException extends RuntimeException {
-    public InvalidFeedbackException() {
+public class FeedbackInvalidException extends RuntimeException {
+    public FeedbackInvalidException() {
         super("The amount of marks is not the same as the length of the word!");
     }
 }
