+++ src/main/java/nl/hu/cisq1/lingo/domain/exception/WordLengthNotSupportedException.java
@@ -1,0 +1,7 @@
+package nl.hu.cisq1.lingo.domain.exception;
+
+public class WordLengthNotSupportedException extends RuntimeException {
+    public WordLengthNotSupportedException(Integer length) {
+        super("Could not find word of length " + length);
+    }
+}
