--- src/main/java/nl/hu/cisq1/lingo/words/domain/exception/WordLengthNotSupportedException.java
@@ -1,7 +1,0 @@
-package nl.hu.cisq1.lingo.words.domain.exception;
-
-public class WordLengthNotSupportedException extends RuntimeException {
-    public WordLengthNotSupportedException(Integer length) {
-        super("Could not find word of length " + length);
-    }
-}
