+++ src/test/java/nl/hu/cisq1/lingo/CiTestConfiguration.java
@@ -1,7 +1,7 @@
 package nl.hu.cisq1.lingo;
 
 import nl.hu.cisq1.lingo.words.WordTestDataFixtures;
-import nl.hu.cisq1.lingo.words.data.SpringWordRepository;
+import nl.hu.cisq1.lingo.data.SpringWordRepository;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
