+++ src/test/java/nl/hu/cisq1/lingo/domain/RoundTest.java
@@ -1,17 +1,17 @@
 package nl.hu.cisq1.lingo.domain;
 
-import nl.hu.cisq1.lingo.words.domain.Word;
 import org.junit.jupiter.api.DisplayName;
-import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Stream;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 class RoundTest {
 
@@ -20,7 +20,7 @@ class RoundTest {
     void showFirstLetterOfWord() {
         Word word = new Word("woord");
 
-        Round round = new Round(1,  word, new ArrayList<>());
+        Round round = new Round(1, word,  new ArrayList<>());
 
         assertEquals(Arrays.asList('w', '_', '_', '_', '_'), round.startRound());
     }
