+++ src/test/java/nl/hu/cisq1/lingo/domain/WordTest.java
@@ -1,0 +1,17 @@
+package nl.hu.cisq1.lingo.domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+class WordTest {
+
+    @Test
+    @DisplayName("length is based on given word")
+    void lengthBasedOnWord() {
+        Word word = new Word("woord");
+        int length = word.getLength();
+        assertEquals(5, length);
+    }
+}
